Setup TypeScript
### Summary

This diff made it possible to write TypeScript in cureview's codebase. To enable it, I had to bump meteor version, and install `adornis:typescript` to add TS support in meteor.

### Test Plan

To show typescript setup works, I migrated two small entry point files to typescript. Everything still works!

### Notes

The meteor official docs recommend `barbatus:typescript`, but it actually doesn't work and the repo of `barbatus:typescript` clearly says it's deprecated, which shows the increasing risk of staying on the platform.